requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alembic==1.8.1
    # via -r requirements.in
anyio==3.6.1
    # via starlette
asyncpg==0.26.0
    # via
    #   -r requirements.in
    #   databases
atomicwrites==1.4.1
    # via pytest
attrs==22.1.0
    # via pytest
bcrypt==3.2.2
    # via passlib
cffi==1.15.1
    # via
    #   bcrypt
    #   cryptography
click==8.1.3
    # via uvicorn
colorama==0.4.5
    # via
    #   click
    #   pytest
cryptography==37.0.4
    # via python-jose
databases[postgresql]==0.6.0
    # via -r requirements.in
ecdsa==0.18.0
    # via python-jose
fastapi==0.79.0
    # via
    #   -r requirements.in
    #   fastapi-jwt-auth
fastapi-jwt-auth==0.5.0
    # via -r requirements.in
greenlet==1.1.2
    # via sqlalchemy
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
iniconfig==1.1.1
    # via pytest
jose==1.0.0
    # via -r requirements.in
mako==1.2.1
    # via alembic
markupsafe==2.1.1
    # via mako
packaging==21.3
    # via pytest
passlib[bcrypt]==1.7.4
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
pyjwt==1.7.1
    # via fastapi-jwt-auth
pyparsing==3.0.9
    # via packaging
pytest==7.1.2
    # via -r requirements.in
python-dotenv==0.20.0
    # via -r requirements.in
python-jose[cryptography]==3.3.0
    # via -r requirements.in
rsa==4.9
    # via python-jose
six==1.16.0
    # via ecdsa
sniffio==1.2.0
    # via anyio
sqlalchemy==1.4.39
    # via
    #   -r requirements.in
    #   alembic
    #   databases
starlette==0.19.1
    # via
    #   -r requirements.in
    #   fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.3.0
    # via
    #   pydantic
    #   starlette
uvicorn==0.18.2
    # via -r requirements.in
psycopg2-binary